Clase valores añadida
